requirements.txt
pandas>=1.4
numpy>=1.21
sentence-transformers>=2.2
scikit-learn>=1.1
torch>=1.11
timm>=0.6
Pillow>=9.2
matplotlib>=3.4
tqdm>=4.64
